Fix data provider issues in WSL ubuntu

+++ src/data_provider/src/sysInfoLinux.cpp
@@ -187,19 +187,19 @@ void SysInfo::getMemory(nlohmann::json& info) const
     std::map<std::string, std::string> systemInfo;
     getSystemInfo(WM_SYS_MEM_DIR, ":", systemInfo);
 
-    auto memTotal{ 1 };
-    auto memFree{ 0 };
+    auto memTotal{ 1ull };
+    auto memFree{ 0ull };
 
     const auto& itTotal { systemInfo.find("MemTotal") };
     if (itTotal != systemInfo.end())
     {
-        memTotal = std::stoi(itTotal->second);
+        memTotal = std::stoull(itTotal->second);
     }
 
     const auto& itFree { systemInfo.find("MemFree") };
     if (itFree != systemInfo.end())
     {
-        memFree = std::stoi(itFree->second);
+        memFree = std::stoull(itFree->second);
     }
 
     const auto ramTotal { memTotal == 0 ? 1 : memTotal };
